+++ mtg/src/main/scala/com/github/gchudnov/mtg/internal/AlgStatic.scala
@@ -3,10 +3,7 @@ package com.github.gchudnov.mtg.internal
 import com.github.gchudnov.mtg.Mark
 import com.github.gchudnov.mtg.Interval
 import com.github.gchudnov.mtg.Domain
-import com.github.gchudnov.mtg.internal.alg.Complement
-import com.github.gchudnov.mtg.internal.alg.Group
-import com.github.gchudnov.mtg.internal.alg.Split
-import com.github.gchudnov.mtg.internal.alg.Intersection
+import com.github.gchudnov.mtg.internal.alg.*
 
 /**
  * Static Interval Operations
@@ -15,48 +12,48 @@ private[mtg] transparent trait AlgStatic:
 
   /**
    * Intersection
+   * 
+   * @see
+   *   [[Intersection.intersection]]
    */
   final def intersection[T: Domain](a: Interval[T], b: Interval[T]): Interval[T] =
     Intersection.intersection(a, b)
 
   /**
    * Span
+   * 
+   * @see
+   *   [[Span.span]]
    */
   final def span[T: Domain](a: Interval[T], b: Interval[T]): Interval[T] =
-    a.span(b)
+    Span.span(a, b)
 
   /**
    * Union
+   * 
+   * @see
+   *   [[Union.union]]
    */
   final def union[T: Domain](a: Interval[T], b: Interval[T]): Interval[T] =
-    a.union(b)
+    Union.union(a, b)
 
   /**
    * Gap
+   * 
+   * @see
+   *   [[Gap.gap]]
    */
   final def gap[T: Domain](a: Interval[T], b: Interval[T]): Interval[T] =
-    a.gap(b)
+    Gap.gap(a, b)
 
   /**
    * Minus
    *
-   * Unlike `a.minus(b)` that returns an interval and not defined when `a.contains(b)`, this method returns a non-empty collection of intervals and can be used when `a.contains(b)`
-   * is true.
-   *
-   * {{{
-   *   a - b = [c1, c2]                       | [a-, pred(b-)], [succ(b+), a+]
-   *
-   *   [**********************************]   | [1,15]  : a
-   *             [************]               | [5,10]  : b
-   *   [*******]                              | [1,4]   : c1
-   *                            [*********]   | [11,15] : c2
-   * --+-------+-+------------+-+---------+-- |
-   *   1       4 5           10          15   |
-   * }}}
+   * @see
+   *   [[Minus.minus]]
    */
   final def minus[T: Domain](a: Interval[T], b: Interval[T]): List[Interval[T]] =
-    if a.contains(b) then List(Interval.make(a.left, b.left.pred), Interval.make(b.right.succ, a.right))
-    else List(a.minus(b))
+    Minus.minus(a, b)
 
   /**
    * Group
